backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
pygithub>=1.58.0
javalang>=0.13.0
python-dotenv>=1.0.0
